+++ woodwork/tests/accessor/test_statistics.py
@@ -23,6 +23,7 @@ from woodwork.logical_types import (
 )
 from woodwork.statistics_utils import (
     _get_describe_dict,
+    _get_mode,
     _make_categorical_for_mutual_info,
     _replace_nans_for_mutual_info
 )
@@ -35,6 +36,27 @@ from woodwork.utils import import_or_none
 
 dd = import_or_none('dask.dataframe')
 ks = import_or_none('databricks.koalas')
+
+
+def test_get_mode():
+    series_list = [
+        pd.Series([1, 2, 3, 4, 2, 2, 3]),
+        pd.Series(['a', 'b', 'b', 'c', 'b']),
+        pd.Series([3, 2, 3, 2]),
+        pd.Series([np.nan, np.nan, np.nan]),
+        pd.Series([pd.NA, pd.NA, pd.NA]),
+        pd.Series([1, 2, np.nan, 2, np.nan, 3, 2]),
+        pd.Series([1, 2, pd.NA, 2, pd.NA, 3, 2])
+    ]
+
+    answer_list = [2, 'b', 2, None, None, 2, 2]
+
+    for series, answer in zip(series_list, answer_list):
+        mode = _get_mode(series)
+        if answer is None:
+            assert mode is None
+        else:
+            assert mode == answer
 
 
 def test_accessor_replace_nans_for_mutual_info():
